Now using JSON for configuration file
Switched from using a plain text file to a JSON file. This has resulted
in less code being needed to access the settings.

Also added function to update the last scan time. This is called if
do_scan is returned True, i.e. after a scan for new podcasts is
authorised.